wrap notifications in partial class
